add theme-launcher cli
